agent_runtime/requirements.txt
anthropic>=0.28.0
pydantic>=2.7.0
httpx>=0.27.0
jsonschema>=4.22.0
redis>=5.0.0

# File tools
openpyxl>=3.1.0
python-docx>=1.1.0
pypdf>=4.2.0

# Observability
langfuse>=2.0.0,<3.0.0
